+++ plugins/action/aaa_services_id_info.py
@@ -30,8 +30,8 @@ argument_spec = dnac_argument_spec()
 argument_spec.update(
     dict(
         id=dict(type="str"),
-        startTime=dict(type="float"),
-        endTime=dict(type="float"),
+        startTime=dict(type="int"),
+        endTime=dict(type="int"),
         headers=dict(type="dict"),
     )
 )
