+++ plugins/action/assurance_get_sensor_test_results_info.py
@@ -30,8 +30,8 @@ argument_spec = dnac_argument_spec()
 argument_spec.update(
     dict(
         siteId=dict(type="str"),
-        startTime=dict(type="float"),
-        endTime=dict(type="float"),
+        startTime=dict(type="int"),
+        endTime=dict(type="int"),
         testFailureBy=dict(type="str"),
         headers=dict(type="dict"),
     )
