+++ plugins/action/field_notices_results_network_devices_count_info.py
@@ -31,7 +31,7 @@ argument_spec.update(
     dict(
         networkDeviceId=dict(type="str"),
         scanStatus=dict(type="str"),
-        noticeCount=dict(type="float"),
+        noticeCount=dict(type="int"),
         headers=dict(type="dict"),
     )
 )
