+++ plugins/action/field_notices_results_notices_info.py
@@ -30,10 +30,10 @@ argument_spec = dnac_argument_spec()
 argument_spec.update(
     dict(
         id=dict(type="str"),
-        deviceCount=dict(type="float"),
+        deviceCount=dict(type="int"),
         type=dict(type="str"),
-        offset=dict(type="float"),
-        limit=dict(type="float"),
+        offset=dict(type="int"),
+        limit=dict(type="int"),
         sortBy=dict(type="str"),
         order=dict(type="str"),
         headers=dict(type="dict"),
